requirements.txt
#discord-related 
discord.py 
git+https://github.com/Rapptz/discord-ext-menus 
#important stuff
python-dotenv 
u-msgpack-python 
aiohttp 
aiohttp-session
aiohttp-jinja2
pygithub 
#less important
timeago 
number-parser 
dateparser 
pillow 
numpy 
pytimeparse 
pytz 
humanize
lxml
psutil